routing working with landing page and profile page (blank except for navbar) and scheduling page, which is broken but it's located correctly!
